+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Fixing NewDecimalField example.
 * Creating FollyUtils.colorHex(Color color) method.
-* Updating FollyUtils.colorParser to accept default color and int color.
+* Updating FollyUtils.colorParser to return null instead default color.
 * Renaming HomeCard backgroundColor to color.
 
 ## [1.2.2] - 2023-07-15